requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==5.0.1
django-celery-beat==2.1.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.4
environs==9.5.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
marshmallow==3.20.2
openapi-codec==1.3.2
packaging==23.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.5
pytz==2024.1
redis==4.4.2
requests==2.28.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
setuptools
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.2.0
